+++ DependencyInjection/CraueConfigExtension.php
@@ -4,15 +4,15 @@ namespace Craue\ConfigBundle\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
 /**
  * Registration of the extension via DI.
  *
  * @author Christian Raue <christian.raue@gmail.com>
- * @copyright 2011-2013 Christian Raue
- * @license http://www.opensource.org/licenses/mit-license.php MIT License
+ * @copyright 2011-2016 Christian Raue
+ * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class CraueConfigExtension extends Extension {
 
@@ -21,6 +21,11 @@ class CraueConfigExtension extends Extension {
 	 */
 	public function load(array $config, ContainerBuilder $container) {
 		$loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+
+		if (!method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix')) {
+			$loader->load('form_legacy.xml'); // for symfony/form < 2.8
+		}
+
 		$loader->load('twig.xml');
 		$loader->load('util.xml');
 	}