+++ Entity/Setting.php
@@ -7,8 +7,8 @@ use Symfony\Component\Validator\Constraints as Assert;
 
 /**
  * @author Christian Raue <christian.raue@gmail.com>
- * @copyright 2011-2013 Christian Raue
- * @license http://www.opensource.org/licenses/mit-license.php MIT License
+ * @copyright 2011-2016 Christian Raue
+ * @license http://opensource.org/licenses/mit-license.php MIT License
  *
  * @ORM\Entity
  * @ORM\Table(name="craue_config_setting")
@@ -24,13 +24,13 @@ class Setting {
 	protected $name;
 
 	/**
-	 * @var string
+	 * @var string|null
 	 * @ORM\Column(name="value", type="string", nullable=true)
 	 */
 	protected $value;
 
 	/**
-	 * @var string
+	 * @var string|null
 	 * @ORM\Column(name="section", type="string", nullable=true)
 	 */
 	protected $section;