+++ Twig/Extension/ConfigTemplateExtension.php
@@ -2,24 +2,37 @@
 
 namespace Craue\ConfigBundle\Twig\Extension;
 
+use Craue\ConfigBundle\Util\Config;
+
 /**
  * @author Christian Raue <christian.raue@gmail.com>
- * @copyright 2011-2013 Christian Raue
- * @license http://www.opensource.org/licenses/mit-license.php MIT License
+ * @copyright 2011-2016 Christian Raue
+ * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class ConfigTemplateExtension extends \Twig_Extension {
 
 	/**
 	 * @var string[]
 	 */
-	protected $sectionOrder;
+	protected $sectionOrder = array();
 
 	/**
-	 * Sets the order in which sections will be rendered.
-	 * @param string[] $sectionOrder
+	 * @var Config
+	 */
+	protected $config;
+
+	/**
+	 * @param string[] $sectionOrder The order in which sections will be rendered.
 	 */
 	public function setSectionOrder(array $sectionOrder = array()) {
 		$this->sectionOrder = $sectionOrder;
+	}
+
+	/**
+	 * @param Config $config
+	 */
+	public function setConfig(Config $config) {
+		$this->config = $config;
 	}
 
 	/**
@@ -33,8 +46,29 @@ class ConfigTemplateExtension extends \Twig_Extension {
 	 * {@inheritDoc}
 	 */
 	public function getFilters() {
+		if (version_compare(\Twig_Environment::VERSION, '1.12', '<')) {
+			return array(
+				'craue_sortSections' => new \Twig_Filter_Method($this, 'sortSections'),
+			);
+		}
+
 		return array(
-			'craue_sortSections' => new \Twig_Filter_Method($this, 'sortSections'),
+			new \Twig_SimpleFilter('craue_sortSections', array($this, 'sortSections')),
+		);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function getFunctions() {
+		if (version_compare(\Twig_Environment::VERSION, '1.12', '<')) {
+			return array(
+				'craue_setting' => new \Twig_Function_Method($this, 'getSetting'),
+			);
+		}
+
+		return array(
+			new \Twig_SimpleFunction('craue_setting', array($this, 'getSetting')),
 		);
 	}
 
@@ -63,6 +97,15 @@ class ConfigTemplateExtension extends \Twig_Extension {
 		}
 
 		return $finalSectionOrder;
+	}
+
+	/**
+	 * @param string $name Name of the setting.
+	 * @return string|null Value of the setting.
+	 * @throws \RuntimeException If the setting is not defined.
+	 */
+	public function getSetting($name) {
+		return $this->config->get($name);
 	}
 
 }